chore: minor plugin refactoring #9228
