+++ jans-casa/app/src/main/resources/labels/user.properties
@@ -10,7 +10,8 @@ usr.main_intro=Your {0} account can be protected by some of the most advanced on
 usr.method.preferred_short=Set a preferred method
 usr.method.preferred_title=Set your preferred authentication method
 usr.method.preferred=Choose the type of credential you want to be prompted first when second factor authentication takes place:
-usr.method.preferred_note=Note some methods may not be available on mobile browsers and an alternative method will be picked instead.
+usr.method.preferred_note= 
+#Note some methods may not be available on mobile browsers and an alternative method will be picked instead.
 
 #2FA
 usr.mfa_title=Second Factor Authentication
