+++ jans-casa/plugins/samples/sample-cred/src/main/java/io/jans/casa/plugins/sampleauthn/SampleCredentialPlugin.java
@@ -1,4 +1,4 @@
-package io.jans.casa.plugins;
+package io.jans.casa.plugins.sampleauthn;
 
 import org.pf4j.Plugin;
 import org.pf4j.PluginWrapper;
